PlaceSearchedリファクタ
